Final cleanups
Rerun the Python 2 compatibility code detector and remove accidentally
added code. Rerun the linters.

Fix reorganizing of test files. There was a problem with the `test_util`
and `utils` file in the `test` module. They have now been merged in the
`test/utils.py` file and all imports adapted.
