+++ setup.py
@@ -33,6 +33,8 @@ setup(
         'tornado>=4.1, <5.0',
         'pyyaml~=5.1.2',
         'pika>=1.0.0',
+        'aio-pika',
+        'aiocontextvars',
         'kiwipy[rmq]~=0.5.5',
     ],
     extras_require={
