+++ test/persistence/test_inmemory.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 import plumpy
-from test.test_utils import ProcessWithCheckpoint
-from test.utils import TestCaseWithLoop
+from test.utils import ProcessWithCheckpoint, TestCaseWithLoop
 
 
 class TestInMemoryPersister(TestCaseWithLoop):
