+++ test/process/__init__.py
@@ -1,6 +1,6 @@
+# -*- coding: utf-8 -*-
 """Plumpy tests"""
 
-from __future__ import absolute_import
 import logging
 import tempfile
 import os
