+++ test/process/test_processes.py
@@ -15,28 +15,28 @@ from .. import utils
 class ForgetToCallParent(plumpy.Process):
 
     def __init__(self, forget_on):
-        super(ForgetToCallParent, self).__init__()
+        super().__init__()
         self.forget_on = forget_on
 
     def on_create(self):
         if self.forget_on != 'create':
-            super(ForgetToCallParent, self).on_create()
+            super().on_create()
 
     def on_run(self):
         if self.forget_on != 'run':
-            super(ForgetToCallParent, self).on_run()
+            super().on_run()
 
     def on_except(self, exc_info):
         if self.forget_on != 'except':
-            super(ForgetToCallParent, self).on_except(exc_info)
+            super().on_except(exc_info)
 
     def on_finish(self, result, successful):
         if self.forget_on != 'finish':
-            super(ForgetToCallParent, self).on_finish(result, successful)
+            super().on_finish(result, successful)
 
     def on_kill(self, msg):
         if self.forget_on != 'kill':
-            super(ForgetToCallParent, self).on_kill(msg)
+            super().on_kill(msg)
 
 
 def test_spec():
@@ -65,7 +65,7 @@ def test_dynamic_inputs():
 
         @classmethod
         def define(cls, spec):
-            super(WithDynamic, cls).define(spec)
+            super().define(spec)
             spec.inputs.dynamic = True
 
     with pytest.raises(ValueError):
@@ -81,7 +81,7 @@ def test_inputs():
 
         @classmethod
         def define(cls, spec):
-            super(Proc, cls).define(spec)
+            super().define(spec)
             spec.input('a')
 
     proc = Proc({'a': 5})
@@ -98,7 +98,7 @@ def test_inputs_default():
 
         @classmethod
         def define(cls, spec):
-            super(Proc, cls).define(spec)
+            super().define(spec)
             spec.input('input', default=5, required=False)
 
     # Supply a value
@@ -117,7 +117,7 @@ def test_inputs_default_that_evaluate_to_false():  #pylint: disable=invalid-name
 
             @classmethod
             def define(cls, spec):
-                super(Proc, cls).define(spec)
+                super().define(spec)
                 spec.input('input', default=def_val)  # pylint: disable=cell-var-from-loop
 
         # Don't supply, use default
@@ -133,7 +133,7 @@ def test_nested_namespace_defaults():
 
         @classmethod
         def define(cls, spec):
-            super(SomeProcess, cls).define(spec)
+            super().define(spec)
             spec.input_namespace('namespace', required=False)
             spec.input('namespace.sub', default=True)
 
@@ -149,7 +149,7 @@ def test_raise_in_define():
 
         @classmethod
         def define(cls, spec):
-            super(BrokenProcess, cls).define(spec)
+            super().define(spec)
             raise ValueError
 
     with pytest.raises(ValueError):
@@ -222,7 +222,7 @@ def test_get_description():
 
         @classmethod
         def define(cls, spec):
-            super(ProcWithSpec, cls).define(spec)
+            super().define(spec)
             spec.input('a', default=1)
 
     for proc_class in utils.TEST_PROCESSES:
@@ -461,7 +461,7 @@ def test_unsuccessful_result():
 
         @classmethod
         def define(cls, spec):
-            super(Proc, cls).define(spec)
+            super().define(spec)
 
         def run(self):
             return plumpy.UnsuccessfulResult(ERROR_CODE)
@@ -592,7 +592,7 @@ class SavePauseProc(plumpy.Process):
     steps_ran = None
 
     def init(self):
-        super(SavePauseProc, self).init()
+        super().init()
         self.steps_ran = []
 
     def run(self):
@@ -616,7 +616,7 @@ class TestProcessNamespace(unittest.TestCase):
 
             @classmethod
             def define(cls, spec):
-                super(NameSpacedProcess, cls).define(spec)
+                super().define(spec)
                 spec.input('some.name.space.a', valid_type=int)
 
         proc = NameSpacedProcess(inputs={'some': {'name': {'space': {'a': 5}}}})
@@ -641,7 +641,7 @@ class TestProcessNamespace(unittest.TestCase):
 
             @classmethod
             def define(cls, spec):
-                super(NameSpacedProcess, cls).define(spec)
+                super().define(spec)
                 spec.input('some.name.space.a', valid_type=int)
                 spec.input('test', valid_type=int, default=6)
                 spec.input('label', valid_type=str, required=False)
@@ -667,7 +667,7 @@ class TestProcessNamespace(unittest.TestCase):
 
             @classmethod
             def define(cls, spec):
-                super(DummyDynamicProcess, cls).define(spec)
+                super().define(spec)
                 spec.input_namespace(namespace)
                 spec.inputs['name']['space'].dynamic = True
                 spec.inputs['name']['space'].valid_type = int
@@ -700,7 +700,7 @@ class TestProcessNamespace(unittest.TestCase):
 
             @classmethod
             def define(cls, spec):
-                super(DummyDynamicProcess, cls).define(spec)
+                super().define(spec)
                 spec.input('output_mode', valid_type=OutputMode, default=OutputMode.NONE)
                 spec.output('required_bool', valid_type=bool)
                 spec.output_namespace(namespace, valid_type=int, dynamic=True)
