+++ test/rmq/test_communicator.py
@@ -13,9 +13,7 @@ from tornado import testing, ioloop
 
 import plumpy
 from plumpy import communications, process_comms
-from test import test_utils
 from test.utils import AsyncTestCase
-from six.moves import range
 
 AWAIT_TIMEOUT = testing.get_async_test_timeout()
 
@@ -50,12 +48,12 @@ class TestLoopCommunicator(CommunicatorTestCase):
     """Make sure the loop communicator is working as expected"""
 
     def setUp(self):
-        super(TestLoopCommunicator, self).setUp()
+        super().setUp()
 
     def tearDown(self):
         # Close the connector before calling super because it will close the loop
         self.rmq_communicator.stop()
-        super(TestLoopCommunicator, self).tearDown()
+        super().tearDown()
 
     @pytest.mark.asyncio
     async def test_broadcast(self):
@@ -127,26 +125,26 @@ class TestTaskActions(CommunicatorTestCase):
     @pytest.mark.asyncio
     async def test_launch(self):
         # Let the process run to the end
-        result = await self.process_controller.launch_process(test_utils.DummyProcess)
+        result = await self.process_controller.launch_process(utils.DummyProcess)
         # Check that we got a result
-        self.assertDictEqual(test_utils.DummyProcess.EXPECTED_OUTPUTS, result)
+        self.assertDictEqual(utils.DummyProcess.EXPECTED_OUTPUTS, result)
 
     @pytest.mark.asyncio
     async def test_launch_nowait(self):
         """ Testing launching but don't wait, just get the pid """
-        pid = await self.process_controller.launch_process(test_utils.DummyProcess, nowait=True)
+        pid = await self.process_controller.launch_process(utils.DummyProcess, nowait=True)
         self.assertIsInstance(pid, uuid.UUID)
 
     @pytest.mark.asyncio
     async def test_execute_action(self):
         """ Test the process execute action """
-        result = await self.process_controller.execute_process(test_utils.DummyProcessWithOutput)
-        self.assertEqual(test_utils.DummyProcessWithOutput.EXPECTED_OUTPUTS, result)
+        result = await self.process_controller.execute_process(utils.DummyProcessWithOutput)
+        self.assertEqual(utils.DummyProcessWithOutput.EXPECTED_OUTPUTS, result)
 
     @pytest.mark.asyncio
     async def test_execute_action_nowait(self):
         """ Test the process execute action """
-        pid = await self.process_controller.execute_process(test_utils.DummyProcessWithOutput, nowait=True)
+        pid = await self.process_controller.execute_process(utils.DummyProcessWithOutput, nowait=True)
         self.assertIsInstance(pid, uuid.UUID)
 
     @pytest.mark.asyncio
@@ -156,7 +154,7 @@ class TestTaskActions(CommunicatorTestCase):
 
         launch_futures = []
         for _ in range(num_to_launch):
-            launch = self.process_controller.launch_process(test_utils.DummyProcess, nowait=True)
+            launch = self.process_controller.launch_process(utils.DummyProcess, nowait=True)
             launch_futures.append(launch)
 
         results = await asyncio.gather(*launch_futures)
@@ -166,11 +164,11 @@ class TestTaskActions(CommunicatorTestCase):
     @pytest.mark.asyncio
     async def test_continue(self):
         """ Test continuing a saved process """
-        process = test_utils.DummyProcessWithOutput()
+        process = utils.DummyProcessWithOutput()
         self.persister.save_checkpoint(process)
         pid = process.pid
         del process
 
         # Let the process run to the end
         result = await self.process_controller.continue_process(pid)
-        self.assertEqual(result, test_utils.DummyProcessWithOutput.EXPECTED_OUTPUTS)
+        self.assertEqual(result, utils.DummyProcessWithOutput.EXPECTED_OUTPUTS)
