+++ test/rmq/test_process_thread_controller.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 import asyncio
 import pytest
