+++ test/test_expose.py
@@ -2,7 +2,6 @@
 from plumpy.ports import PortNamespace
 from plumpy.processes import Process
 from plumpy.process_spec import ProcessSpec
-from test.test_utils import NewLoopProcess
 from test import utils
 
 
@@ -14,7 +13,7 @@ class TestExposeProcess(utils.TestCaseWithLoop):
         def validator_function(input, port):
             pass
 
-        class BaseNamespaceProcess(NewLoopProcess):
+        class BaseNamespaceProcess(utils.NewLoopProcess):
 
             @classmethod
             def define(cls, spec):
@@ -25,7 +24,7 @@ class TestExposeProcess(utils.TestCaseWithLoop):
                 spec.inputs['namespace'].valid_type = (int, float)
                 spec.inputs['namespace'].validator = validator_function
 
-        class BaseProcess(NewLoopProcess):
+        class BaseProcess(utils.NewLoopProcess):
 
             @classmethod
             def define(cls, spec):
@@ -35,7 +34,7 @@ class TestExposeProcess(utils.TestCaseWithLoop):
                 spec.inputs.dynamic = True
                 spec.inputs.valid_type = str
 
-        class ExposeProcess(NewLoopProcess):
+        class ExposeProcess(utils.NewLoopProcess):
 
             @classmethod
             def define(cls, spec):
@@ -148,7 +147,7 @@ class TestExposeProcess(utils.TestCaseWithLoop):
         """Test that the exclude argument of exposed_inputs works correctly and excludes ports from being absorbed."""
         BaseProcess = self.BaseProcess
 
-        class ExcludeProcess(NewLoopProcess):
+        class ExcludeProcess(utils.NewLoopProcess):
 
             @classmethod
             def define(cls, spec):
@@ -166,7 +165,7 @@ class TestExposeProcess(utils.TestCaseWithLoop):
         """Test that the include argument of exposed_inputs works correctly and includes only specified ports."""
         BaseProcess = self.BaseProcess
 
-        class ExcludeProcess(NewLoopProcess):
+        class ExcludeProcess(utils.NewLoopProcess):
 
             @classmethod
             def define(cls, spec):
@@ -184,7 +183,7 @@ class TestExposeProcess(utils.TestCaseWithLoop):
         """Test that passing both exclude and include raises."""
         BaseProcess = self.BaseProcess
 
-        class ExcludeProcess(NewLoopProcess):
+        class ExcludeProcess(utils.NewLoopProcess):
 
             @classmethod
             def define(cls, spec):
@@ -371,7 +370,7 @@ class TestExposeProcess(utils.TestCaseWithLoop):
         """Test the include rules can be nested and are properly unwrapped."""
         BaseNamespaceProcess = self.BaseNamespaceProcess
 
-        class ExposeProcess(NewLoopProcess):
+        class ExposeProcess(utils.NewLoopProcess):
 
             @classmethod
             def define(cls, spec):
@@ -385,7 +384,7 @@ class TestExposeProcess(utils.TestCaseWithLoop):
         """Test the include rules can be nested and are properly unwrapped."""
         BaseNamespaceProcess = self.BaseNamespaceProcess
 
-        class ExposeProcess(NewLoopProcess):
+        class ExposeProcess(utils.NewLoopProcess):
 
             @classmethod
             def define(cls, spec):
@@ -401,7 +400,7 @@ class TestExposeProcess(utils.TestCaseWithLoop):
         """Test the include rules can be nested and are properly unwrapped."""
         BaseNamespaceProcess = self.BaseNamespaceProcess
 
-        class ExposeProcess(NewLoopProcess):
+        class ExposeProcess(utils.NewLoopProcess):
 
             @classmethod
             def define(cls, spec):
@@ -417,7 +416,7 @@ class TestExposeProcess(utils.TestCaseWithLoop):
         """Test the include rules can be nested and are properly unwrapped."""
         BaseNamespaceProcess = self.BaseNamespaceProcess
 
-        class ExposeProcess(NewLoopProcess):
+        class ExposeProcess(utils.NewLoopProcess):
 
             @classmethod
             def define(cls, spec):
@@ -433,7 +432,7 @@ class TestExposeProcess(utils.TestCaseWithLoop):
         """Test the exclude rules can be nested and are properly unwrapped."""
         BaseNamespaceProcess = self.BaseNamespaceProcess
 
-        class ExposeProcess(NewLoopProcess):
+        class ExposeProcess(utils.NewLoopProcess):
 
             @classmethod
             def define(cls, spec):
@@ -449,7 +448,7 @@ class TestExposeProcess(utils.TestCaseWithLoop):
         """Test the exclude rules can be nested and are properly unwrapped."""
         BaseNamespaceProcess = self.BaseNamespaceProcess
 
-        class ExposeProcess(NewLoopProcess):
+        class ExposeProcess(utils.NewLoopProcess):
 
             @classmethod
             def define(cls, spec):
@@ -463,7 +462,7 @@ class TestExposeProcess(utils.TestCaseWithLoop):
         """Test the exclude rules can be nested and are properly unwrapped."""
         BaseNamespaceProcess = self.BaseNamespaceProcess
 
-        class ExposeProcess(NewLoopProcess):
+        class ExposeProcess(utils.NewLoopProcess):
 
             @classmethod
             def define(cls, spec):
@@ -479,7 +478,7 @@ class TestExposeProcess(utils.TestCaseWithLoop):
         """Test the exclude rules can be nested and are properly unwrapped."""
         BaseNamespaceProcess = self.BaseNamespaceProcess
 
-        class ExposeProcess(NewLoopProcess):
+        class ExposeProcess(utils.NewLoopProcess):
 
             @classmethod
             def define(cls, spec):
